Wildlife healthcare system
Star wildlife company has created a website called Wildlife healthcare system for Sri Lanka. 
Sri Lanka is a country with a rich wildlife tourism industry. Therefore, the health of the wildlife 
living in the official forest parks of Sri Lanka has a great impact on Sri Lanka. There, wildlife 
is prone to frequent accidents and illnesses, and in such cases, it is important to treat wildlife 
that have been exposed to those hazards and diseases. It is therefore important that the 
treatment is done quickly and properly. In this way, the staff of the wildlife offices provide 
treatment to the wildlife when they need it. When wildlife is treated in this way, there are 
shortages and problems with some wildlife medicine and surgical equipment. These 
deficiencies and problems are mainly due to the lack of wildlife medicine in the treatment 
areas for the general population. Therefore, the fact that they are restricted to zoos only 
means that only wildlife veterinarians have the legal right to treat wildlife. Therefore, this 
website has been created by linking several parks in Sri Lanka with such wildlife medicine 
stores and treatment centers. Therefore, through this website,